mv ead
